Remove cast to int to prevent rounding errors during conversion to pixels.

+++ core/JavaRendererUtils/src/main/java/ArmyC2/C2SD/Utilities/PointConversion.java
@@ -164,10 +164,10 @@ public class PointConversion implements IPointConversion {
         }
         temp = (calcValue / _pixelMultiplierX);
 
-        pixel.x = (int)temp;
+        pixel.x = temp;
 
         temp = ((_geoTop - coord.getY()) / _pixelMultiplierY);//yMultiplier);
-        pixel.y = (int)temp;
+        pixel.y = temp;
 
         return pixel;
     }
